Extract thread memory artifact semantics
