+++ codex-rs/context-maintenance-policy/templates/thread_memory/schema.json
